+++ scripts/version-bump.js
@@ -8,6 +8,16 @@ replace({
 })
   .then(changedFiles => {
     console.log('Modified files:', changedFiles.join(', '));
+    return replace({
+      files: ['plugin.xml'],
+      // from: /id="@sentry\/cordova" version="\d+\.\d+.\d+"/g,
+      // to: `id="@sentry/cordova" version="${pjson.version}"`,
+      from: /id="sentry-cordova" version="\d+\.\d+.\d+"/g,
+      to: `id="sentry-cordova" version="${pjson.version}"`,
+    });
+  })
+  .then(changedFiles => {
+    console.log('Modified files:', changedFiles.join(', '));
   })
   .catch(error => {
     console.error('Error occurred:', error);